Move potential common code to module_utils / plugin_utils.
